+++ src/es_utils/promises.rs
@@ -1,10 +1,12 @@
-use mozjs::rust::jsapi_wrapped::NewPromiseObject;
-use mozjs::rust::HandleObject;
-
 use crate::es_utils::objects::{get_constructor, get_es_obj_prop_val_as_string};
+use crate::es_utils::{report_es_ex, EsErrorInfo};
 use mozjs::jsapi::JSContext;
 use mozjs::jsapi::JSObject;
 use mozjs::jsval::NullValue;
+use mozjs::rust::jsapi_wrapped::NewPromiseObject;
+use mozjs::rust::jsapi_wrapped::RejectPromise;
+use mozjs::rust::jsapi_wrapped::ResolvePromise;
+use mozjs::rust::{HandleObject, HandleValue};
 
 pub fn object_is_promise(context: *mut JSContext, obj: HandleObject) -> bool {
     // todo this is not the best way of doing this, we need to get the promise object of the global scope and see if that is the same as the objects constructor
@@ -26,13 +28,57 @@ pub fn object_is_promise(context: *mut JSContext, obj: HandleObject) -> bool {
 pub fn new_promise(context: *mut JSContext) -> *mut JSObject {
     // second is executor
 
-    rooted!(in(context) let null = NullValue().to_object());
+    rooted!(in(context) let null = NullValue().to_object_or_null());
     let null_handle: HandleObject = null.handle();
     unsafe { NewPromiseObject(context, null_handle) }
 }
 
 pub fn new_promise_with_exe(context: *mut JSContext, executor: HandleObject) -> *mut JSObject {
     unsafe { NewPromiseObject(context, executor) }
+}
+
+pub fn resolve_promise(
+    context: *mut JSContext,
+    promise: HandleObject,
+    resolution_value: HandleValue,
+) -> Result<(), EsErrorInfo> {
+    let ok = unsafe { ResolvePromise(context, promise, resolution_value) };
+    if ok {
+        Ok(())
+    } else {
+        if let Some(err) = report_es_ex(context) {
+            Err(err)
+        } else {
+            Err(EsErrorInfo {
+                message: "unknown error resolving promise".to_string(),
+                filename: "".to_string(),
+                lineno: 0,
+                column: 0,
+            })
+        }
+    }
+}
+
+pub fn reject_promise(
+    context: *mut JSContext,
+    promise: HandleObject,
+    rejection_value: HandleValue,
+) -> Result<(), EsErrorInfo> {
+    let ok = unsafe { RejectPromise(context, promise, rejection_value) };
+    if ok {
+        Ok(())
+    } else {
+        if let Some(err) = report_es_ex(context) {
+            Err(err)
+        } else {
+            Err(EsErrorInfo {
+                message: "unknown error rejecting promise".to_string(),
+                filename: "".to_string(),
+                lineno: 0,
+                column: 0,
+            })
+        }
+    }
 }
 
 #[cfg(test)]